% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_impute.R
\name{do_impute}
\alias{do_impute}
\title{Sampling multiple copies of imputed values and cell factors at given parameters.}
\usage{
do_impute(dat, Y, beta, lambda, sigma, mu, pi, pos_mean = NULL,
  pos_sd = NULL, celltype = NULL, mcmc = 10, burnin = 2,
  verbose = F, pg = 0.5, cutoff = 0.1)
}
\arguments{
\item{dat}{scRNASeq data matrix. Each row is a gene, each column is a cell.}

\item{Y}{Initial imputed data, output from \emph{SIMPLE_B} or \emph{SIMPLE}.}

\item{beta}{Factor loadings, output from \emph{SIMPLE_B} or \emph{SIMPLE}.}

\item{lambda}{Variances of factors for each cluster, output from \emph{SIMPLE_B} or \emph{SIMPLE}.}

\item{sigma}{Variances of idiosyncratic noises for each cluster, output from \emph{SIMPLE_B} or \emph{SIMPLE}.}

\item{mu}{Mean expression for each cluster, output from \emph{SIMPLE_B} or \emph{SIMPLE}.}

\item{pi}{Probabilites of cells belong to each cluster, output from \emph{SIMPLE_B} or \emph{SIMPLE}.}

\item{pos_mean}{Gene mean. If centerized each gene before estimating the parameters, provide the overall mean of gene expression removed from the data matrix, output from \emph{SIMPLE_B} or \emph{SIMPLE}.
Default is NULL.}

\item{pos_sd}{Gene standard deviation. If scaled each gene before estimating the parameters, provide the overall standard deviation of gene expression removed from the data matrix, output from \emph{SIMPLE_B} or \emph{SIMPLE}. Default is NULL.
Default is NULL.}
}
\value{
\code{do_impute} returns a list of imputation results in the following order.
\enumerate{
  \item{loglik}{The log-likelihood of the imputed gene expression at each iteration.}
  \item{impt}{A matrix contains the expectation of imputed expression.}
 \item{impt_var}{A matrix contains the variance of each imputed entry.}
  \item{EF}{Posterior means of factors}
  \item{varF}{Posterior covariance matrix of factors}
 \item{consensus_cluster}{Score for the clustering stability of each cell by multiple imputations. }
}
}
\description{
\code{do_impute} get the posterior mean and variance for the imputed values and factors given the parameters, which can be returned from \emph{SIMPLE} or \emph{SIMPLE_B}.
}
\author{
Zhirui Hu, \email{zhiruihu@g.harvard.edu}

Songpeng Zu, \email{songpengzu@g.harvard.edu}
}
