% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMPLE.R
\name{init_impute}
\alias{init_impute}
\title{Initialize imputation for each individual gene}
\usage{
init_impute(Y2, M0, clus, p_min = 0.6, cutoff = 0.1, verbose = F)
}
\arguments{
\item{Y2}{scRNASeq data matrix. Each row is a gene, each column is a cell.}

\item{M0}{number of cell types}

\item{clus}{A numeric vector for the cell type labels of cells in the scRNASeq. The labels must start from 1 to the number of types (M0).}

\item{p_min}{Restrict the max dropout rate to be 1-p_min. Default = 0.6.}

\item{cutoff}{The value below cutoff is treated as no expression. Default = 0.1.}

\item{verbose}{Whether to plot some intermediate result. Default = False.}
}
\value{
Imputed gene expression matrix, treat each gene independently
}
\description{
Initialize imputation for each individual gene
}
\details{
Fit each gene expression by a zero-inflated censored Gaussian distribution and return a random sample of imputed values for initialization
}
\author{
Zhirui Hu, \email{zhiruihu@g.harvard.edu}

Songpeng Zu, \email{songpengzu@g.harvard.edu}
}
