% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMPLE_B.R
\name{init_impute_bulk}
\alias{init_impute_bulk}
\title{Initialize imputation for each individual gene with known dropout rate for each cell type}
\usage{
init_impute_bulk(Y2, clus, bulk, pg1, cutoff = 0.1, verbose = F)
}
\arguments{
\item{Y2}{scRNASeq data matrix. Each row is a gene, each column is a cell.}

\item{clus}{A numeric vector for the cell type labels of cells in the scRNASeq. The labels must start from 1 to the number of types.}

\item{bulk}{Gene expression matrix of bulk RNASeq, only used for plot if verbose = T.}

\item{pg1}{lower bound of Dropout rate (at least 1- pg1) for each cell type.}

\item{cutoff}{The value below cutoff is treated as no expression. Default = 0.1.}

\item{verbose}{Whether to plot some intermediate result. Default = False.}
}
\value{
Imputed gene expression matrix, treat each gene independently
}
\description{
Initialize imputation for each individual gene with known dropout rate for each cell type
}
\details{
Fit each gene expression by a zero-inflated censored Gaussian distribution and return a random sample of imputed values for initialization
}
\author{
Zhirui Hu, \email{zhiruihu@g.harvard.edu}

Songpeng Zu, \email{songpengzu@g.harvard.edu}
}
