% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMPLE.R
\name{init_impute}
\alias{init_impute}
\title{Initialize imputed matrix integrating bulk RNASeq}
\usage{
init_impute(Y2, M0, clus, p_min = 0.6, cutoff = 0.5, verbose = F)
}
\arguments{
\item{Y2}{scRNASeq data matrix. Each row is a gene, each column is a cell.}

\item{clus}{A numeric vector for the cell type labels of cells in the scRNASeq. Each cell type has corresponding the mean expression in the bulk RNASeq data. The labels must start from 1 to the number of types. See details.}

\item{cutoff}{The value below cutoff is treated as no expression. Default = 0.1.}

\item{verbose}{Whether to plot some intermediate result. Default = False.}

\item{bulk}{Bulk RNASeq data matrix. Each row is a gene which must be ordered the same as the scRNASeq data. Each column is a cell type which must be ordered as the cell type label in \emph{clus}.}

\item{pg1}{A data matrix for the dropout rate for each gene which is the ratio between mean expression in the scRNASeq and bulk RNASeq. Each row is a gene which must be ordered the same as the scRNASeq data. Each column is a cell type which must be ordered as the cell type label in \emph{clus}.}
}
\value{
Imputed gene expression matrix
}
\description{
Initialize imputed matrix integrating bulk RNASeq
}
\details{
If the scRNASeq data does not have the cell type information, then all cells
  can be treated as a single type and the input bulk RNASeq should also have
  one column that is the mean expression over all cell types.
}
\author{
Zhirui Hu, \email{zhiruihu@g.harvard.edu}

Songpeng Zu, \email{songpengzu@g.harvard.edu}
}
